Move unit_conversion_exception down to UnitsHelper
