+++ src/bmi/Bmi_Adapter.cpp
@@ -2,6 +2,7 @@
 #include "bmi/State_Exception.hpp"
 #include "utilities/FileChecker.h"
 #include "utilities/logging_utils.h"
+#include "core/mediator/UnitsHelper.hpp"
 
 namespace models {
 namespace bmi {
